Added SoundSystem.java and split Sound.java in SoundPlayer.java which is implemented in SimpleSoundPlayer.java and MusicSoundPlayer.java
