+++ src/main/java/de/athalion/game/twodgame/sound/SoundSystem.java
@@ -1,0 +1,6 @@
+package de.athalion.game.twodgame.sound;
+
+public class SoundSystem {
+
+
+}
