+++ src/main/java/de/athalion/game/twodgame/sound/Sounds.java
@@ -1,0 +1,29 @@
+package de.athalion.game.twodgame.sound;
+
+public class Sounds {
+
+    // MUSIC
+    public static String MUSIC_MENU = "/sound/music/menu.wav";
+    public static String MUSIC_DEMON_BATTLE_INTRO = "/sound/music/demon_battle/demon_battle_intro.wav";
+    public static String MUSIC_DEMON_BATTLE_LOOP = "/sound/music/demon_battle/demon_battle_loop.wav";
+    public static String MUSIC_DEMON_BATTLE_OUTRO = "/sound/music/demon_battle/demon_battle_outro.wav";
+    public static String MUSIC_DEMON_BATTLE_SMALL = "/sound/music/demon_battle/demon_battle_small.wav";
+
+    // EFFECTS
+    public static String EFFECT_COIN = "/sound/coin.wav";
+    public static String EFFECT_POWERUP = "/sound/powerup.wav";
+    public static String EFFECT_UNLOCK = "/sound/unlock.wav";
+    public static String EFFECT_FANFARE = "/sound/fanfare.wav";
+    public static String EFFECT_HIT_MONSTER = "/sound/hitmonster.wav";
+    public static String EFFECT_RECEIVE_DAMAGE = "/sound/receivedamage.wav";
+    public static String EFFECT_SWING_WEAPON = "/sound/swingweapon.wav";
+    public static String EFFECT_LEVEL_UP = "/sound/levelup.wav";
+    public static String EFFECT_CURSOR = "/sound/effects/cursor.wav";
+    public static String EFFECT_TEXT = "/sound/text.wav";
+    public static String EFFECT_LOGO = "/sound/logo.wav";
+
+    // ENVIRONMENT
+    public static String ENVIRONMENT_RAIN = "/sound/environment/rain.wav";
+    public static String ENVIRONMENT_RAIN_DULL = "/sound/environment/rain_dull.wav";
+
+}
